Extract setup wizard skills step presenter
