+++ Packages/src/Editor/Presentation/Setup/SetupWizardSkillsStepPresenter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f22074d3a0cc4d2895895bc150236085
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
